backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
spoon-ai-sdk
watchdog
aiohttp
yfinance
pandas
numpy
websockets
